Started lisp/robot-name.
Need to find better alternative to 'loop' macro, I don't
need to collect values, just want to repeat until a condition
is met (find a name that's not already taken).

Also check on using funcall with lambdas. It might be easier
to just define more functions, but it would be good to know
how to use lambda functions.

I started writing an object but I'm not actually sure how it
works so I'm not sure if I'm doing it right yet. I bookmarked
a page on Common Lisp user-defined objects, so hopefully that
continues to be helpful.