Get callbacks to work in rust/react
The callbacks are working. Unfortunately, I have to have unwrap and panic in a
lot of places. I think I want to go back to having separate `HashMap`s for
input and compute cells.

In addition, I need to clean up the logic for caching values and updating the
cached values on updates.

In addition, I should return the cached value instead of re-evaluating the
function when querying the value of compute cells.