Finish haskell/go-counting
This is pretty good except for the fact that it re-creates a territory
for every empty space and then merges them using a set. It would be
better to keep track of territories and only create a new territory if
the current cell is not already in an existing territory.