Optimize alphametics
My alphametics was inefficient because it stored each character in the
addend columns individually. We don't really care about the individual
characters or the order of characters, but simply the number of each
kind of character. We now use a `Counter` to store the count for each
character.

This improves the long test case
(`test_puzzle_with_ten_letters_and_199_addends`) to be solved in
roughly 13 seconds on my machine. That is still rather slow but is at
least workable.